CacheSettings: Tip podataka ponovo izmenjen uint -> ushort, zbog potencijalnih hardverskih ogranicenja.

+++ GitHub_API/CacheSettings.cs
@@ -2,9 +2,9 @@
 
 public static class CacheSettings
 {
-    public static uint MaxEntries { get; private set; } = 1000;
+    public static ushort MaxEntries { get; private set; } = 1000;
     public static TimeSpan CleanupPeriod { get; private set; } = TimeSpan.FromMinutes(30);
-    public static uint MaxEntryContributorCount { get; private set; } = 1000;
+    public static ushort MaxEntryContributorCount { get; private set; } = 1000;
     public static bool CachingEnabled { get; private set; } = true;
 
 }
